tracemyflow/requirements.txt

Django==4.2.9
pyyaml==6.0.1
psycopg2-binary==2.9.9
whitenoise==6.6.0
gunicorn==21.2.0
python-dotenv==1.0.0